+++ Sources/PhotoEditorKit/Editor/Common/Protocols/EditorProtocols.swift
@@ -36,7 +36,7 @@ protocol TextShadow {
 
 protocol TextFont {
     func font(_ : UIFont)
-    func font(size: Double)
+    func font(size: Int)
 }
 
 protocol TextColor {
@@ -48,5 +48,5 @@ protocol TextFade {
 }
 
 protocol TextLayerRotate {
-    func layerRotate(angle: Double)
+    func layerRotate(angle: Int)
 }
